+++ src/en/anineko/src/eu/kanade/tachiyomi/animeextension/en/anineko/AniNeko.kt
@@ -1,7 +1,6 @@
 package eu.kanade.tachiyomi.animeextension.en.anineko
 
 import android.net.Uri
-import android.util.Base64
 import androidx.preference.PreferenceScreen
 import aniyomi.lib.doodextractor.DoodExtractor
 import aniyomi.lib.playlistutils.PlaylistUtils
@@ -24,20 +23,10 @@ import keiyoushi.utils.useAsJsoup
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.Request
 import okhttp3.Response
-import java.net.ServerSocket
-import java.net.Socket
-import java.util.concurrent.Executors
-import kotlin.math.min
 
 class AniNeko :
     AnimeHttpSource(),
     ConfigurableAnimeSource {
-
-    private data class VideoInfo(
-        val quality: String?,
-        val audioType: String?,
-        val bandwidth: Long?,
-    )
 
     override val name = "AniNeko"
 
@@ -56,20 +45,22 @@ class AniNeko :
 
     private val playlistUtils by lazy { PlaylistUtils(client, headers) }
 
-    // ============================== Popular / Latest ==============================
+    // ============================= Popular ==============================
 
-    override fun popularAnimeRequest(page: Int): Request = GET("$baseUrl/browser?page=$page", headers)
+    override fun popularAnimeRequest(page: Int): Request = GET("$baseUrl/browse?sort=release_date&page=$page", headers)
 
     override fun popularAnimeParse(response: Response): AnimesPage = searchAnimeParse(response)
 
-    override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/browser?sort=recently_updated&page=$page", headers)
+    // ============================= Latest ===============================
+
+    override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/browse?sort=recently_updated&page=$page", headers)
 
     override fun latestUpdatesParse(response: Response): AnimesPage = searchAnimeParse(response)
 
     // ============================== Search ==============================
 
     override fun searchAnimeRequest(page: Int, query: String, filters: AnimeFilterList): Request {
-        val urlBuilder = "$baseUrl/browser".toHttpUrl().newBuilder()
+        val urlBuilder = "$baseUrl/browse".toHttpUrl().newBuilder()
         urlBuilder.addQueryParameter("page", page.toString())
 
         if (query.isNotBlank()) {
@@ -78,37 +69,37 @@ class AniNeko :
 
         filters.forEach { filter ->
             when (filter) {
-                is GenreFilter -> {
+                is Filters.GenreFilter -> {
                     filter.getCheckedUriParts().forEach {
                         urlBuilder.addQueryParameter("genre[]", it)
                     }
                 }
 
-                is TypeFilter -> {
+                is Filters.TypeFilter -> {
                     filter.getCheckedUriParts().forEach {
                         urlBuilder.addQueryParameter("type[]", it)
                     }
                 }
 
-                is StatusFilter -> {
+                is Filters.StatusFilter -> {
                     filter.getCheckedUriParts().forEach {
                         urlBuilder.addQueryParameter("status[]", it)
                     }
                 }
 
-                is LanguageFilter -> {
+                is Filters.LanguageFilter -> {
                     filter.getCheckedUriParts().forEach {
                         urlBuilder.addQueryParameter("language[]", it)
                     }
                 }
 
-                is YearFilter -> {
+                is Filters.YearFilter -> {
                     filter.getCheckedUriParts().forEach {
                         urlBuilder.addQueryParameter("year[]", it)
                     }
                 }
 
-                is SortFilter -> {
+                is Filters.SortFilter -> {
                     if (!filter.isDefault()) {
                         urlBuilder.addQueryParameter("sort", filter.toUriPart())
                     }
@@ -140,14 +131,43 @@ class AniNeko :
         return AnimesPage(animes, hasNextPage)
     }
 
-    // ============================== Anime Details ==============================
+    // ============================= Filters ==============================
+
+    open class UriPartFilter(displayName: String, private val vals: Array<Pair<String, String>>) : AnimeFilter.Select<String>(displayName, vals.map { it.first }.toTypedArray()) {
+        fun toUriPart() = vals[state].second
+        fun isDefault() = state == 0
+    }
+
+    private class CheckBoxVal(name: String, state: Boolean = false) : AnimeFilter.CheckBox(name, state)
+
+    open class CheckBoxFilterList(name: String, val vals: Array<Pair<String, String>>) : AnimeFilter.Group<AnimeFilter.CheckBox>(name, vals.map { CheckBoxVal(it.first, false) }) {
+        fun getCheckedUriParts(): List<String> = state.mapIndexedNotNull { index, checkbox ->
+            if (checkbox.state) vals[index].second else null
+        }
+    }
+
+    override fun getFilterList(): AnimeFilterList = AnimeFilterList(
+        Filters.SortFilter(),
+        AnimeFilter.Separator(),
+        Filters.GenreFilter(),
+        AnimeFilter.Separator(),
+        Filters.TypeFilter(),
+        AnimeFilter.Separator(),
+        Filters.StatusFilter(),
+        AnimeFilter.Separator(),
+        Filters.LanguageFilter(),
+        AnimeFilter.Separator(),
+        Filters.YearFilter(),
+    )
+
+    // =========================== Anime Details ==========================
 
     override fun animeDetailsRequest(anime: SAnime): Request = GET("$baseUrl${anime.url}", headers)
 
     override fun animeDetailsParse(response: Response): SAnime {
         val document = response.useAsJsoup()
         return SAnime.create().apply {
-            val titleLang = preferences.getString(PREF_TITLE_LANG_KEY, PREF_TITLE_LANG_DEFAULT)!!
+            val titleLang = preferences.getString(TITLE_LANG_KEY, TITLE_LANG_DEFAULT)!!
             val mainTitle = document.selectFirst("h1")?.text() ?: ""
             val altTitle = document.selectFirst("div.nv-info-alt-title")?.text() ?: ""
             title = if (titleLang == "Romaji/Japanese" && altTitle.isNotBlank()) {
@@ -177,7 +197,7 @@ class AniNeko :
         }
     }
 
-    // ============================== Episode List ==============================
+    // =========================== Episode List ===========================
 
     override fun episodeListRequest(anime: SAnime): Request = GET("$baseUrl${anime.url}", headers)
 
@@ -199,7 +219,7 @@ class AniNeko :
         return list.reversed()
     }
 
-    // ============================== Video List ==============================
+    // ============================ Video List ============================
 
     override fun videoListRequest(episode: SEpisode): Request = GET("$baseUrl${episode.url}", headers)
 
@@ -209,7 +229,7 @@ class AniNeko :
 
         val videos = buttons.parallelCatchingFlatMapBlocking { button ->
             val iframeUrl = button.attr("data-video")
-            if (iframeUrl.isBlank()) return@parallelCatchingFlatMapBlocking emptyList<Video>()
+            if (iframeUrl.isBlank()) return@parallelCatchingFlatMapBlocking emptyList()
 
             val serverName = button.ownText().trim()
             val rawType = button.selectFirst("span")?.text() ?: ""
@@ -285,62 +305,57 @@ class AniNeko :
             }
         }
 
-        val excludedServers = preferences.getStringSet(PREF_EXCLUDE_SERVERS_KEY, emptySet()) ?: emptySet()
-        val excludedAudios = preferences.getStringSet(PREF_EXCLUDE_AUDIO_KEY, emptySet()) ?: emptySet()
+        val excludedServers = preferences.getStringSet(EXCLUDE_SERVERS_KEY, emptySet()) ?: emptySet()
+        val excludedAudios = preferences.getStringSet(EXCLUDE_AUDIO_KEY, emptySet()) ?: emptySet()
 
-        val preferredQuality = preferences.getString(PREF_QUALITY_KEY, PREF_QUALITY_DEFAULT)!!
-        val preferredAudioType = preferences.getString(PREF_TYPE_KEY, PREF_TYPE_DEFAULT)!!
-        val preferredHost = preferences.getString(PREF_HOST_KEY, PREF_HOST_DEFAULT)!!
+        val preferredQuality = preferences.getString(QUALITY_KEY, QUALITY_DEFAULT)!!
+        val preferredAudioType = preferences.getString(TYPE_KEY, TYPE_DEFAULT)!!
+        val preferredHost = preferences.getString(HOST_KEY, HOST_DEFAULT)!!
 
-        val parsedVideos = videos
+        val filteredVideos = videos
             .filter { video ->
                 val matchesServer = excludedServers.any { video.quality.contains(it, ignoreCase = true) }
                 val matchesAudio = excludedAudios.any { video.quality.contains(it, ignoreCase = true) }
                 !matchesServer && !matchesAudio
             }
-            .map { video -> video to parseVideoInfo(video.quality) }
-
-        return parsedVideos
-            .sortedWith(
-                compareBy<Pair<Video, VideoInfo>>(
-                    { (_, info) ->
-                        if (info.quality == preferredQuality &&
-                            info.audioType?.equals(preferredAudioType, ignoreCase = true) == true
-                        ) {
-                            0
-                        } else {
-                            1
-                        }
-                    },
-                    { (_, info) -> if (info.quality == preferredQuality) 0 else 1 },
-                    { (_, info) ->
-                        if (info.audioType?.equals(preferredAudioType, ignoreCase = true) == true) 0 else 1
-                    },
-                    { (video, _) -> if (video.quality.startsWith("$preferredHost - ", ignoreCase = true)) 0 else 1 },
-                    { (_, info) ->
-                        qualityOrder.indexOf(info.quality).let { if (it == -1) Int.MAX_VALUE else it }
-                    },
-                    { (_, info) -> -(info.bandwidth ?: -1L) },
-                    { (video, _) -> video.quality },
-                ),
-            )
-            .map { (video, _) -> video }
-    }
-
-    private fun parseVideoInfo(quality: String): VideoInfo {
-        val videoQuality = qualityRegex.find(quality)?.value?.lowercase()
-        val audioType = audioTypes.firstOrNull { quality.contains(it, ignoreCase = true) }
-        val bandwidth = speedRegex.find(quality)?.let { match ->
-            val value = match.groupValues[1].toDoubleOrNull() ?: return@let null
-            val multiplier = when (match.groupValues[2].lowercase()) {
-                "gb" -> 1_000_000_000
-                "mb" -> 1_000_000
-                "kb" -> 1_000
-                else -> 1
+            // Filter out bare "Video" tracks that lack a standard resolution
+            .filterNot { video ->
+                video.quality.contains("Video", ignoreCase = true) &&
+                    QUALITY_ENTRIES.none { video.quality.contains(it, ignoreCase = true) }
             }
-            (value * multiplier).toLong()
+
+        val isDoodPreferred = preferredHost.equals("Doodstream", ignoreCase = true)
+
+        fun isDoodVideo(video: Video): Boolean = video.quality.contains("Doodstream", ignoreCase = true) ||
+            video.quality.contains("dood", ignoreCase = true)
+
+        val qualitiesList = QUALITY_ENTRIES.reversed()
+
+        return if (isDoodPreferred) {
+            val (doodVideos, otherVideos) = filteredVideos.partition { isDoodVideo(it) }
+
+            val sortedDood = doodVideos.sortedWith(
+                compareByDescending<Video> { it.quality.contains(preferredAudioType, true) }
+                    .thenByDescending { video -> qualitiesList.indexOfLast { video.quality.contains(it, true) } }
+                    .thenByDescending { it.quality.contains(preferredHost, true) },
+            )
+
+            val sortedOthers = otherVideos.sortedWith(
+                compareByDescending<Video> { it.quality.contains(preferredQuality, true) }
+                    .thenByDescending { it.quality.contains(preferredAudioType, true) }
+                    .thenByDescending { video -> qualitiesList.indexOfLast { video.quality.contains(it, true) } }
+                    .thenByDescending { it.quality.contains(preferredHost, true) },
+            )
+
+            sortedDood + sortedOthers
+        } else {
+            filteredVideos.sortedWith(
+                compareByDescending<Video> { it.quality.contains(preferredQuality, true) }
+                    .thenByDescending { it.quality.contains(preferredAudioType, true) }
+                    .thenByDescending { video -> qualitiesList.indexOfLast { video.quality.contains(it, true) } }
+                    .thenByDescending { it.quality.contains(preferredHost, true) },
+            )
         }
-        return VideoInfo(videoQuality, audioType, bandwidth)
     }
 
     private fun addServerName(serverName: String, quality: String): String = if (serverName.isBlank() || quality.startsWith("$serverName - ", ignoreCase = true)) {
@@ -349,514 +364,80 @@ class AniNeko :
         "$serverName - $quality"
     }
 
-    // ============================== Preferences ==============================
+    // ============================ Preferences ===========================
 
     override fun setupPreferenceScreen(screen: PreferenceScreen) {
         screen.addListPreference(
-            key = PREF_QUALITY_KEY,
-            title = "Preferred Quality",
-            entries = listOf("1080p", "720p", "480p", "360p"),
-            entryValues = listOf("1080p", "720p", "480p", "360p"),
-            default = PREF_QUALITY_DEFAULT,
+            key = TITLE_LANG_KEY,
+            title = "Preferred Title Language",
+            entries = TITLE_LANG_ENTRIES,
+            entryValues = TITLE_LANG_ENTRIES,
+            default = TITLE_LANG_DEFAULT,
             summary = "%s",
         )
         screen.addListPreference(
-            key = PREF_TYPE_KEY,
-            title = "Preferred Audio Type",
-            entries = listOf("Soft Sub", "Hard Sub", "Dub"),
-            entryValues = listOf("Soft Sub", "Hard Sub", "Dub"),
-            default = PREF_TYPE_DEFAULT,
-            summary = "%s",
-        )
-        screen.addListPreference(
-            key = PREF_HOST_KEY,
+            key = HOST_KEY,
             title = "Preferred Host",
-            entries = listOf("HD-1", "HD-2", "StreamHG", "Earnvids", "Doodstream"),
-            entryValues = listOf("HD-1", "HD-2", "StreamHG", "Earnvids", "Doodstream"),
-            default = PREF_HOST_DEFAULT,
+            entries = HOST_ENTRIES,
+            entryValues = HOST_ENTRIES,
+            default = HOST_DEFAULT,
+            summary = "%s",
+        )
+        screen.addListPreference(
+            key = QUALITY_KEY,
+            title = "Preferred Quality",
+            entries = QUALITY_ENTRIES,
+            entryValues = QUALITY_ENTRIES,
+            default = QUALITY_DEFAULT,
+            summary = "%s",
+        )
+        screen.addListPreference(
+            key = TYPE_KEY,
+            title = "Preferred Audio Type",
+            entries = TYPE_ENTRIES,
+            entryValues = TYPE_ENTRIES,
+            default = TYPE_DEFAULT,
             summary = "%s",
         )
         screen.addSetPreference(
-            key = PREF_EXCLUDE_SERVERS_KEY,
+            key = EXCLUDE_SERVERS_KEY,
             default = emptySet(),
             title = "Exclude Host",
             summary = "Select servers to exclude from the video list",
-            entries = listOf("HD-1", "HD-2", "StreamHG", "Earnvids", "Doodstream"),
-            entryValues = listOf("HD-1", "HD-2", "StreamHG", "Earnvids", "Doodstream"),
+            entries = HOST_ENTRIES,
+            entryValues = HOST_ENTRIES,
         )
         screen.addSetPreference(
-            key = PREF_EXCLUDE_AUDIO_KEY,
+            key = EXCLUDE_AUDIO_KEY,
             default = emptySet(),
             title = "Exclude Audio Types",
             summary = "Select audio formats to exclude from the video list",
-            entries = listOf("Soft Sub", "Hard Sub", "Dub"),
-            entryValues = listOf("Soft Sub", "Hard Sub", "Dub"),
-        )
-        screen.addListPreference(
-            key = PREF_TITLE_LANG_KEY,
-            title = "Preferred Title Language",
-            entries = listOf("English", "Romaji/Japanese"),
-            entryValues = listOf("English", "Romaji/Japanese"),
-            default = PREF_TITLE_LANG_DEFAULT,
-            summary = "%s",
+            entries = TYPE_ENTRIES,
+            entryValues = TYPE_ENTRIES,
         )
     }
-
-    // ============================== Filters ==============================
-
-    open class UriPartFilter(displayName: String, private val vals: Array<Pair<String, String>>) : AnimeFilter.Select<String>(displayName, vals.map { it.first }.toTypedArray()) {
-        fun toUriPart() = vals[state].second
-        fun isDefault() = state == 0
-    }
-
-    private class CheckBoxVal(name: String, state: Boolean = false) : AnimeFilter.CheckBox(name, state)
-
-    open class CheckBoxFilterList(name: String, val vals: Array<Pair<String, String>>) : AnimeFilter.Group<AnimeFilter.CheckBox>(name, vals.map { CheckBoxVal(it.first, false) }) {
-        fun getCheckedUriParts(): List<String> = state.mapIndexedNotNull { index, checkbox ->
-            if (checkbox.state) vals[index].second else null
-        }
-    }
-
-    class GenreFilter : CheckBoxFilterList("Genres", GENRES)
-    class TypeFilter : CheckBoxFilterList("Types", TYPES)
-    class StatusFilter : CheckBoxFilterList("Status", STATUSES)
-    class LanguageFilter : CheckBoxFilterList("Language/Version", LANGUAGES)
-    class YearFilter : CheckBoxFilterList("Years", YEARS)
-    class SortFilter : UriPartFilter("Sort By", SORT_BY)
-
-    override fun getFilterList(): AnimeFilterList = AnimeFilterList(
-        SortFilter(),
-        AnimeFilter.Separator(),
-        GenreFilter(),
-        AnimeFilter.Separator(),
-        TypeFilter(),
-        AnimeFilter.Separator(),
-        StatusFilter(),
-        AnimeFilter.Separator(),
-        LanguageFilter(),
-        AnimeFilter.Separator(),
-        YearFilter(),
-    )
 
     companion object {
-        private const val PREF_QUALITY_KEY = "preferred_quality"
-        private const val PREF_QUALITY_DEFAULT = "1080p"
+        const val QUALITY_KEY = "preferred_quality"
+        const val QUALITY_DEFAULT = "1080p"
+        val QUALITY_ENTRIES = listOf("1080p", "720p", "480p", "360p")
 
-        private const val PREF_TYPE_KEY = "preferred_type"
-        private const val PREF_TYPE_DEFAULT = "Soft Sub"
+        const val TYPE_KEY = "preferred_type"
+        const val TYPE_DEFAULT = "Soft Sub"
+        val TYPE_ENTRIES = listOf("Soft Sub", "Hard Sub", "Dub")
 
-        private const val PREF_HOST_KEY = "preferred_host"
-        private const val PREF_HOST_DEFAULT = "HD-1"
+        const val HOST_KEY = "preferred_host"
+        const val HOST_DEFAULT = "HD-1"
+        val HOST_ENTRIES = listOf("HD-1", "HD-2", "StreamHG", "Earnvids", "Doodstream")
 
-        private const val PREF_EXCLUDE_SERVERS_KEY = "exclude_servers"
-        private const val PREF_EXCLUDE_AUDIO_KEY = "exclude_audio"
+        const val EXCLUDE_SERVERS_KEY = "exclude_servers"
 
-        private const val PREF_TITLE_LANG_KEY = "preferred_title_lang"
-        private const val PREF_TITLE_LANG_DEFAULT = "English"
+        const val EXCLUDE_AUDIO_KEY = "exclude_audio"
+
+        const val TITLE_LANG_KEY = "preferred_title_lang"
+        const val TITLE_LANG_DEFAULT = "English"
+        val TITLE_LANG_ENTRIES = listOf("English", "Romaji/Japanese")
 
         private val vibeRegex = Regex("""const src\s*=\s*"([^"]+)"""")
-        private val qualityRegex = Regex("""(?<!\d)(?:360|480|720|1080)p\b""", RegexOption.IGNORE_CASE)
-        private val speedRegex = Regex("""(\d+(?:\.\d+)?)\s*(GB|MB|KB|bytes?)/s\b""", RegexOption.IGNORE_CASE)
-        private val audioTypes = listOf("Soft Sub", "Hard Sub", "Dub")
-        private val qualityOrder = listOf("1080p", "720p", "480p", "360p")
-
-        private val GENRES = arrayOf(
-            Pair("Action", "action"),
-            Pair("Adventure", "adventure"),
-            Pair("Cars", "cars"),
-            Pair("Comedy", "comedy"),
-            Pair("Dementia", "dementia"),
-            Pair("Demons", "demons"),
-            Pair("Drama", "drama"),
-            Pair("Ecchi", "ecchi"),
-            Pair("Fantasy", "fantasy"),
-            Pair("Game", "game"),
-            Pair("Harem", "harem"),
-            Pair("Historical", "historical"),
-            Pair("Horror", "horror"),
-            Pair("Isekai", "isekai"),
-            Pair("Josei", "josei"),
-            Pair("Kids", "kids"),
-            Pair("Magic", "magic"),
-            Pair("Mahou Shoujo", "mahou-shoujo"),
-            Pair("Martial Arts", "martial-arts"),
-            Pair("Mecha", "mecha"),
-            Pair("Military", "military"),
-            Pair("Music", "music"),
-            Pair("Mystery", "mystery"),
-            Pair("Parody", "parody"),
-            Pair("Police", "police"),
-            Pair("Psychological", "psychological"),
-            Pair("Romance", "romance"),
-            Pair("Samurai", "samurai"),
-            Pair("School", "school"),
-            Pair("Sci-Fi", "sci-fi"),
-            Pair("Seinen", "seinen"),
-            Pair("Shoujo", "shoujo"),
-            Pair("Shoujo Ai", "shoujo-ai"),
-            Pair("Shounen", "shounen"),
-            Pair("Shounen Ai", "shounen-ai"),
-            Pair("Slice of Life", "slice-of-life"),
-            Pair("Space", "space"),
-            Pair("Sports", "sports"),
-            Pair("Super Power", "super-power"),
-            Pair("Supernatural", "supernatural"),
-            Pair("Thriller", "thriller"),
-            Pair("Vampire", "vampire"),
-        )
-
-        private val TYPES = arrayOf(
-            Pair("TV", "1"),
-            Pair("Movie", "2"),
-            Pair("OVA", "3"),
-            Pair("ONA", "4"),
-            Pair("Special", "5"),
-            Pair("Music", "6"),
-            Pair("TV_SHORT", "7"),
-        )
-
-        private val STATUSES = arrayOf(
-            Pair("Ongoing", "Ongoing"),
-            Pair("Completed", "Completed"),
-            Pair("Upcoming", "info"),
-        )
-
-        private val LANGUAGES = arrayOf(
-            Pair("Subbed", "sub"),
-            Pair("Dubbed", "dub"),
-        )
-
-        private val YEARS = (2026 downTo 2000).map { Pair(it.toString(), it.toString()) }.toTypedArray()
-
-        private val SORT_BY = arrayOf(
-            Pair("Latest Update", "recently_updated"),
-            Pair("Release Date", "release_date"),
-            Pair("Recently Added", "recently_added"),
-            Pair("Title A-Z", "title_az"),
-        )
-    }
-}
-
-class LocalProxy(private val client: okhttp3.OkHttpClient) {
-    private var serverSocket: ServerSocket? = null
-    private val executor = Executors.newFixedThreadPool(
-        maxOf(2, Runtime.getRuntime().availableProcessors() * 2),
-    )
-    var port: Int = 0
-        private set
-
-    /** True only when the local proxy server started successfully. */
-    val isAvailable: Boolean get() = port > 0 && serverSocket?.isClosed == false
-
-    init {
-        try {
-            val ss = ServerSocket(0)
-            serverSocket = ss
-            port = ss.localPort
-            executor.execute {
-                while (!ss.isClosed) {
-                    try {
-                        val socket = ss.accept()
-                        socket.soTimeout = 30_000
-                        executor.execute { handleSocket(socket) }
-                    } catch (_: Exception) {}
-                }
-            }
-        } catch (_: Exception) {
-            // proxy server unavailable; isAvailable will return false
-        }
-    }
-
-    /** Stop accepting new connections and release all resources. */
-    fun shutdown() {
-        try {
-            serverSocket?.close()
-        } catch (_: Exception) {}
-        executor.shutdownNow()
-    }
-
-    /**
-     * Returns a local proxy URL for [targetUrl], or [targetUrl] itself when the
-     * proxy server failed to start (port == 0).
-     */
-    fun getProxyUrl(targetUrl: String, headers: okhttp3.Headers?): String {
-        if (!isAvailable) return targetUrl
-        val encodedUrl = Base64.encodeToString(targetUrl.toByteArray(), Base64.URL_SAFE or Base64.NO_WRAP or Base64.NO_PADDING)
-        val headersStr = headers?.let { h ->
-            val sb = StringBuilder()
-            for (i in 0 until h.size) {
-                sb.append(h.name(i)).append(":").append(h.value(i)).append("\n")
-            }
-            sb.toString()
-        } ?: ""
-        val encodedHeaders = Base64.encodeToString(headersStr.toByteArray(), Base64.URL_SAFE or Base64.NO_WRAP or Base64.NO_PADDING)
-        val ext = if (targetUrl.contains(".m3u8") || targetUrl.contains("mpegurl")) "playlist.m3u8" else "segment.ts"
-        return "http://127.0.0.1:$port/proxy/$ext?url=$encodedUrl&headers=$encodedHeaders"
-    }
-
-    private fun handleSocket(socket: Socket) {
-        var requestParsed = false
-        try {
-            val input = socket.getInputStream()
-            val reader = input.bufferedReader()
-            val firstLine = reader.readLine() ?: return
-            val parts = firstLine.split(" ")
-            if (parts.size < 2) return
-            val rawPath = parts[1]
-            val queryIndex = rawPath.indexOf('?')
-            val queryString = if (queryIndex != -1) rawPath.substring(queryIndex) else ""
-            val pathWithoutQuery = if (queryIndex != -1) rawPath.substring(0, queryIndex) else rawPath
-
-            val path = if (pathWithoutQuery.startsWith("http://") || pathWithoutQuery.startsWith("https://")) {
-                Uri.parse(pathWithoutQuery).path ?: ""
-            } else {
-                pathWithoutQuery
-            }
-
-            if (!path.startsWith("/proxy")) {
-                sendError(socket, 404, "Not Found")
-                return
-            }
-
-            val httpUrl = ("http://127.0.0.1$path$queryString").toHttpUrl()
-            val encodedUrl = httpUrl.queryParameter("url")
-            val encodedHeaders = httpUrl.queryParameter("headers") ?: ""
-
-            if (encodedUrl.isNullOrEmpty()) {
-                sendError(socket, 400, "Missing url parameter")
-                return
-            }
-
-            val targetUrl = String(Base64.decode(encodedUrl, Base64.URL_SAFE or Base64.NO_WRAP or Base64.NO_PADDING))
-            val isM3u8Request = targetUrl.contains(".m3u8") || path.contains("playlist.m3u8")
-
-            val targetHeaders = okhttp3.Headers.Builder()
-            if (encodedHeaders.isNotEmpty()) {
-                val headersStr = String(Base64.decode(encodedHeaders, Base64.URL_SAFE or Base64.NO_WRAP or Base64.NO_PADDING))
-                headersStr.split("\n").forEach { line ->
-                    val headerParts = line.split(":", limit = 2)
-                    if (headerParts.size == 2) {
-                        targetHeaders.set(headerParts[0].trim(), headerParts[1].trim())
-                    }
-                }
-            }
-
-            var clientRange: String? = null
-            var line: String?
-            while (reader.readLine().also { line = it } != null) {
-                if (line!!.isEmpty()) break
-                val headerParts = line!!.split(":", limit = 2)
-                if (headerParts.size == 2) {
-                    val name = headerParts[0].trim()
-                    val value = headerParts[1].trim()
-                    if (name.equals("Range", ignoreCase = true) && !isM3u8Request) {
-                        // HD-1 wraps the MPEG-TS payload in a PNG response. The
-                        // player range is relative to the unwrapped payload, so
-                        // it must be applied after the wrapper is removed.
-                        clientRange = value
-                    }
-                }
-            }
-
-            val request = okhttp3.Request.Builder()
-                .url(targetUrl)
-                .headers(targetHeaders.build())
-                .build()
-
-            requestParsed = true
-            client.newCall(request).execute().use { response ->
-                sendResponse(socket, response, targetUrl, encodedHeaders, clientRange)
-            }
-        } catch (e: Exception) {
-            // Only attempt to send an HTTP error response when the request was
-            // fully parsed; otherwise the socket's streams may be in an
-            // inconsistent state and writing to them could throw or produce
-            // garbage on the wire.
-            if (requestParsed) {
-                try {
-                    sendError(socket, 500, e.message ?: "Internal Error")
-                } catch (_: Exception) {}
-            }
-        } finally {
-            try {
-                socket.close()
-            } catch (_: Exception) {}
-        }
-    }
-
-    private fun sendResponse(
-        socket: Socket,
-        response: Response,
-        targetUrl: String,
-        encodedHeaders: String,
-        clientRange: String?,
-    ) {
-        val out = socket.getOutputStream()
-        val isM3u8 = targetUrl.contains(".m3u8") || response.header("Content-Type")?.contains("mpegurl") == true
-
-        if (isM3u8) {
-            val bodyString = response.body.string()
-            val modifiedContent = processM3u8(bodyString, targetUrl, encodedHeaders)
-            val body = modifiedContent.toByteArray()
-            out.write("HTTP/1.1 ${response.code} ${response.message}\r\n".toByteArray())
-            writeForwardedHeaders(out, response, isM3u8 = true)
-            out.write("Content-Length: ${body.size}\r\n".toByteArray())
-            out.write("Content-Type: application/vnd.apple.mpegurl\r\n".toByteArray())
-            out.write("Connection: close\r\n\r\n".toByteArray())
-            out.write(body)
-            out.flush()
-            return
-        }
-
-        val stripped = stripPngHeader(response.body.bytes())
-        val range = parseRange(clientRange, stripped.size)
-        val body = range?.let { stripped.copyOfRange(it.first, it.last + 1) } ?: stripped
-        val status = if (range == null) 200 else 206
-        val contentType = if (isMpegTs(stripped)) {
-            "video/mp2t"
-        } else {
-            response.header("Content-Type") ?: "application/octet-stream"
-        }
-
-        out.write("HTTP/1.1 $status ${if (status == 206) "Partial Content" else "OK"}\r\n".toByteArray())
-        writeForwardedHeaders(out, response, isM3u8 = false)
-        if (range != null) {
-            out.write("Accept-Ranges: bytes\r\n".toByteArray())
-            out.write("Content-Range: bytes ${range.first}-${range.last}/${stripped.size}\r\n".toByteArray())
-        }
-        out.write("Content-Length: ${body.size}\r\n".toByteArray())
-        out.write("Content-Type: $contentType\r\n".toByteArray())
-        out.write("Connection: close\r\n\r\n".toByteArray())
-        out.write(body)
-        out.flush()
-    }
-
-    private fun writeForwardedHeaders(
-        out: java.io.OutputStream,
-        response: Response,
-        isM3u8: Boolean,
-    ) {
-        val headers = response.headers
-        for (i in 0 until headers.size) {
-            val name = headers.name(i)
-            val value = headers.value(i)
-            if (name.equals("Connection", ignoreCase = true) ||
-                name.equals("Transfer-Encoding", ignoreCase = true) ||
-                name.equals("Content-Type", ignoreCase = true) ||
-                name.equals("Content-Length", ignoreCase = true) ||
-                (!isM3u8 && name.equals("Content-Range", ignoreCase = true)) ||
-                (!isM3u8 && name.equals("Accept-Ranges", ignoreCase = true))
-            ) {
-                continue
-            }
-            out.write("$name: $value\r\n".toByteArray())
-        }
-    }
-
-    private fun parseRange(rangeHeader: String?, size: Int): IntRange? {
-        if (rangeHeader.isNullOrBlank() || size <= 0 || !rangeHeader.startsWith("bytes=")) return null
-
-        val range = rangeHeader.removePrefix("bytes=").substringBefore(",")
-        val parts = range.split("-", limit = 2)
-        if (parts.size != 2) return null
-
-        val start = parts[0].toIntOrNull()
-        val end = parts[1].toIntOrNull()
-        return when {
-            start != null && start >= 0 && start < size -> {
-                start..min(end ?: (size - 1), size - 1)
-            }
-            start == null && end != null && end > 0 -> {
-                maxOf(0, size - end)..(size - 1)
-            }
-            else -> null
-        }
-    }
-
-    private fun processM3u8(content: String, playlistUrl: String, encodedHeaders: String): String {
-        val lines = content.split(Regex("""\r?\n"""))
-        val builder = StringBuilder(content.length * 2)
-
-        for (line in lines) {
-            val trimmed = line.trim()
-            if (trimmed.isEmpty()) {
-                builder.append("\n")
-                continue
-            }
-
-            if (trimmed.startsWith("#")) {
-                if (trimmed.startsWith("#EXT-X-KEY") || trimmed.startsWith("#EXT-X-MAP") || trimmed.startsWith("#EXT-X-MEDIA")) {
-                    val uriRegex = Regex("""URI=["']?([^"',\s>]+)["']?""")
-                    uriRegex.find(trimmed)?.let { match ->
-                        val uriValue = match.groupValues[1]
-                        val resolvedUri = resolveUrl(playlistUrl, uriValue)
-                        val proxiedUri = getProxyUrlWithEncodedHeaders(resolvedUri, encodedHeaders)
-                        builder.append(trimmed.replace(uriValue, proxiedUri))
-                    } ?: builder.append(trimmed)
-                } else {
-                    builder.append(trimmed)
-                }
-            } else {
-                val resolvedUri = resolveUrl(playlistUrl, trimmed)
-                builder.append(getProxyUrlWithEncodedHeaders(resolvedUri, encodedHeaders))
-            }
-            builder.append("\n")
-        }
-
-        return builder.toString()
-    }
-
-    private fun getProxyUrlWithEncodedHeaders(targetUrl: String, encodedHeaders: String): String {
-        val encodedUrl = Base64.encodeToString(targetUrl.toByteArray(), Base64.URL_SAFE or Base64.NO_WRAP or Base64.NO_PADDING)
-        val ext = if (targetUrl.contains(".m3u8") || targetUrl.contains("mpegurl")) "playlist.m3u8" else "segment.ts"
-        return "http://127.0.0.1:$port/proxy/$ext?url=$encodedUrl&headers=$encodedHeaders"
-    }
-
-    private fun resolveUrl(baseUrl: String, relativeUrl: String): String = try {
-        baseUrl.toHttpUrl().resolve(relativeUrl)?.toString() ?: relativeUrl
-    } catch (_: Exception) {
-        relativeUrl
-    }
-
-    private fun stripPngHeader(data: ByteArray): ByteArray {
-        if (data.size < 8) return data
-        val isPng = data[0] == (-119).toByte() && data[1] == 80.toByte() && data[2] == 78.toByte() && data[3] == 71.toByte()
-        if (!isPng) return data
-        var videoStart = -1
-        val length = data.size - 4
-        for (i in 0 until length) {
-            if (data[i] == 73.toByte() && data[i + 1] == 69.toByte() && data[i + 2] == 78.toByte() && data[i + 3] == 68.toByte()) {
-                videoStart = i + 8
-                break
-            }
-        }
-        if (videoStart < 0 || videoStart >= data.size) return data
-        val tsData = data.copyOfRange(videoStart, data.size)
-        val iMin = min(tsData.size - 188, 400)
-        for (offset in 0 until iMin) {
-            if (tsData[offset] == 0x47.toByte() && tsData[offset + 188] == 0x47.toByte()) {
-                return tsData.copyOfRange(offset, tsData.size)
-            }
-        }
-        return tsData
-    }
-
-    private fun isMpegTs(data: ByteArray): Boolean {
-        if (data.size < 376) return false
-        val maxOffset = min(data.size - 376, 400)
-        return (0..maxOffset).any { offset ->
-            data[offset] == 0x47.toByte() &&
-                data[offset + 188] == 0x47.toByte() &&
-                data[offset + 376] == 0x47.toByte()
-        }
-    }
-
-    private fun sendError(socket: Socket, code: Int, message: String) {
-        val out = socket.getOutputStream()
-        out.write("HTTP/1.1 $code $message\r\n".toByteArray())
-        out.write("Content-Type: text/plain\r\n".toByteArray())
-        out.write("\r\n".toByteArray())
-        out.write(message.toByteArray())
-        out.flush()
     }
 }
